Add system_reset() function to Mbed OS
