+++ drivers/CAN.cpp
@@ -18,7 +18,7 @@
 #if DEVICE_CAN
 
 #include "cmsis.h"
-#include "platform/mbed_sleep.h"
+#include "platform/mbed_power_mgmt.h"
 
 namespace mbed {
